requirements.txt
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
Bootstrap-Flask==2.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
customtkinter==5.2.2
darkdetect==0.8.0
flasgger==0.9.7.1
Flask==3.0.0
Flask-WTF==1.2.1
frozendict==2.3.10
gunicorn==21.2.0
html5lib==1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lxml==4.9.3
MarkupSafe==2.1.3
mistune==3.0.2
multitasking==0.0.11
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pandas-ta==0.3.14b0
peewee==3.17.0
pillow==10.2.0
pybase64==1.3.2
python-dateutil==2.8.2
Python-IO==0.3
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.2
yfinance==0.2.33
